Move unit tests to separate spec file Simplify index.ts by not importing filename constants Fix failing test after packageNmae -> depName
